tests/requirements.txt
pytest==8.3.5
pytest-playwright==0.6.2
